backend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-graphql-jwt==0.4.0
djangorestframework==3.14.0
Faker==12.0.1
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
mixer==7.2.2
promise==2.3
psycopg2-binary==2.9.7
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
